When foldopen contains "search", expand folds the cursor is already inside
For example:

```python
if cond:
    f()    #
    g()    # these three lines are folded
    h()    #
```

Displaying something like:

```python
 if cond:
+--  3 lines: f()    #
```

If the cursor is within the fold and the user types `/g`, vim reports no match (the same for subsequent `n` / `N` presses).

With this change, the fold is expanded and the line containing `g()` is visible. This is only done if `'foldopen'` contains `"search"`