xxd: Fix -R argument parsing
`-R` argument is specified as optional, yet, as written, it always
consumes an argument.  Or was it intended to be a required value and the
documentation need to be changed from `[WHEN]` to `WHEN`?

"auto" value has no effect, meaning if `-R` is repeated and the second
case uses "auto" it does not overwrite the previous setting.