runtime(vim): Improve Vim9 and legacy-script comment highlighting
This is a simple attempt to only highlight Vim9 and legacy-script
comments in the appropriate syntactic contexts.

Vim9-script comments are highlighted at top level in a Vim9-script file,
in all :def functions, and in all :au and :com command blocks.
Legacy-script comments are highlighted at top level in a legacy script
file and in all :func functions.

Fixes #13047, #11307 and #9587.