Extend guard for sound support on OS X
Fixes build on legacy versions where required coreaudio functionality
may not be available.
NSSoundDelegate apparently was introduced in Snow Leopard yet the
build breaks on it. Guarding off enabling sound support to El Capitan
as that's the next version I had access to for testing (it may work on
earlier versions)
https://developer.apple.com/documentation/appkit/nssounddelegate
Vim builds on OS X Tiger 10.4 and newer with this change.