Add extended attribute support (Linux)
It's been a long standing issue, that if you write a file with extended
attributes and backupcopy=no, the file will loose the extended
attributes.

So this patch adds support for retrieving the extended attributes and
copying it to the new file. It currently only works on linux, mainly I
don't know the different APIs for other systems (BSD, MacOSX and
Solaris).  On linux, this should be supported since Kernel 2.4 or
something, so this should be pretty safe to use now.

I think it makes sense to have this enabled in normal builds so it is
not so surprising if one loses those attributes, but not sure.

I also added it explicitly to the :version output as well as make it
able to check using `:echo has("xattr")`, to have users easily check
that this is available.

In contrast to the similar support for SELINUX and SMACK support (which
also internally uses extended attributes), I have made this a FEAT_XATTR
define, instead of the similar HAVE_XATTR, but that should just be a
implementation detail (but happy to change if wanted).

todo:
[X] try to write tests.

fixes: #306

Signed-off-by: Christian Brabandt <cb@256bit.org>