Add completion handlers for simple enum and flag list string options
'whichwrap': This option is quite weird as it's a comma-separated flag
list (the commas are actually optional). Fix up the logic for set-= to
handle it properly, and remove the ',' from WW_ALL so auto-complete
works properly.

'concealcursor': This option is a flag list option but wasn't tagged
correctly. Make sure to tag it as a P_FLAGLIST to make sure it follows
the same logic as other similar options (e.g. 'mouse').