make :0cq return zero in exmode
Currently, when an error occurs in ex-mode (e.g. -e or -E) and an error
occurs, '-0cq' will still abort with an error.

Let's make it possible to return zero even in that case, so that Vim
explicitly exits with zero exit code, even if an error occurred.

Add a few tests to verify this.

This also allows to use '&&' in the test_crash test suite.