restricted mode: disable allow setting of environment variables
This may be used to escape the restricted mode and execute some payload
using GCONV_PATH when calling the iconv() function (by using the
`iconv_open()` function.

Disable setting the GCONV_PATH environment variable, even so this is no
complete protection, since if that environment variable is already in
the environment, it may still execute shell commands.

We could disable the iconv() function completely in restricted mode, not
sure if this is worth it?