[use-after-free] in ex_substitute
There was a report that Vim was using heap-use-after free in a deeply
recursive :substitute call chain.

The following commit will fix this. I haven't been able to create a
reproducible test case yet unfortunately.

Until now, we used the sub variable mostly as a pointer to the stack,
however, sometimes it was actually allocated, so in addition the
variable sub_copy was used to remember if there was anything to free.

However, when calling :s recursively, it may happen that sub points to
the allocated variable old_sub, which might be freed later in a
recursive call, which means that sub would then point to freed memory.

So let's just always allocate memory for sub but of course now we need
to make sure to properly free the memory whenever we leave the function.

Since sub is now always allocated, we don't need the sub_copy variable
anymore, which makes it slightly more easier to read.

TODO: add a proper test case that causes the initial issue.

Signed-off-by: Christian Brabandt <cb@256bit.org>